requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiohttp>=3.9.1
openai-whisper>=20231117
numpy>=1.24.3
soundfile>=0.12.1
scipy>=1.11.4
python-dotenv>=1.0.0
ollama>=0.1.0
# Torch is required for Whisper
torch>=2.0.0
